vs2022
compile success and can play only in debug mode. exclude bullet physics ....

+++ leven/src/materials.h
@@ -79,7 +79,7 @@ public:
 
 	unsigned int size() const
 	{
-		return materials_.size();
+		return (int)materials_.size();
 	}
 
 	// ----------------------------------------------------------------------------
